Moved short URL detail reducer to short-urls module
