+++ src/visits/ShortUrlVisits.tsx
@@ -4,9 +4,9 @@ import { boundToMercureHub } from '../mercure/helpers/boundToMercureHub';
 import { ShlinkVisitsParams } from '../api/types';
 import { parseQuery } from '../utils/helpers/query';
 import { Topics } from '../mercure/helpers/Topics';
+import { ShortUrlDetail } from '../short-urls/reducers/shortUrlDetail';
 import { ShortUrlVisits as ShortUrlVisitsState } from './reducers/shortUrlVisits';
 import ShortUrlVisitsHeader from './ShortUrlVisitsHeader';
-import { ShortUrlDetail } from './reducers/shortUrlDetail';
 import VisitsStats from './VisitsStats';
 
 export interface ShortUrlVisitsProps extends RouteComponentProps<{ shortCode: string }> {
